requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
pillow>=10.0.0
plotly>=5.22.0
pandas>=2.0.0
ultralytics>=8.0.0
easyocr>=1.7.1
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0